Adjustment so Tox works with Python3
The 2to3 tool runs against the code and dumps the updated code into the
build directory. Because of this, 'tests' was added to packages in the setup.py
file. This change moves tests under SoftLayer and removes it from the setup.py.
This change also removes Python 2.4 and 2.5 from the tox file.