Clarify CCI transaction output
* Rename the "provisioning" column to "active_transaction"
It's slightly more accurate, and it's just a preference of mine not to
introduce new terminology. If users interact with the API in other ways,
it'll be called a transaction, right? Or is the terminology changing
globally?

* Display "<None>" for an empty active transaction
The empty string means that transaction would always have to be the last
column, because you couldn't count on awk column offsets in selecting
individual fields. Also, it makes it hard to grep for CCI that _don't_
currently have an active transaction.