Adds Message Queue Service Manager/Connection Class
This was a bit of a difficult one because message queue is an altogether different
API and don't support all the same ways of authenticating as the core SoftLayer API
can and certainly supports username/API key auth differently.

The Message Queue CLI module has been refactored to use the new code and is probably
a good reference.