Always provided HTTP[S] proxy to `requests`
Earlier, only the protocol of the provided proxy was used to forward it
to `requests` calls. So, if proxy with `ftp` or non-http[s] provided it
wouldn't get used while making API calls. This caused us to use provided proxy as HTTP[S] proxy
regardless of the protocol in the proxy URL itself.