Fixes a failing CLI setup test.
This test is failing because of two PRs interacting in unexpected ways.
One change changes the tests to use a real XML-RPC transport against a real HTTP
server which responds with stubbed data. The other change fixes a bug where the endpoint
being used to validate credentials was simply using the default.

Now, the tests be sure to use a stubbed transport so that changing the endpoint will not
cause `slcli setup` to use the default endpoint and make real API calls against it.