Add security group functionality
This adds security group API helpers to the network manager. This allows
users to perform CRUD operations on security groups, security group
rules, and interface attachments to security groups. CLI commands were
also added to perform these API operations on the command line.

Arguments were also added to the instance creation to allow VSIs to be
booted with security groups on the public and/or private interfaces of
the newly created VSI.

Security group information was added to the VSI detail CLI, so the
security groups attached to the interfaces of a VSI are output as part
of the detail.