Set formatting.blank() properly
Because some of the .get() calls return an empty string, they're not
actually getting filled with formatting.blank(). This changes the get()
calls to add the or after the get() call. This way, if the returned
value is empty string (so not set), we'll fill it with a
formatting.blank() instead of filling it with nothing.