Update needed fixes for CLI and add doc
The CLI appears to fail out if it gets AttributeError as part of any of
the manager calls, so in order to not make it look like a
softlayer-pythyon but, we need to spit out something from
SoftLayer.exceptions. The base exceptions.SoftLayerError was used.

Presets also needed to be handle differently. When listing the presets,
both the activePresets and accountRestrictedActivePresets need to be
retrieved and merged together before being returned.

Finally, extra doc was added to verify_order() and place_order() to
make it easier to understand how to use it.