Fix parsing logic in volume-detail commands to handle invalid types
Some certain types of transactions which are currently returned in
a volume's activeTransactions contain transactionStatus sub-properties
of inconsistent types. There is a plan to fix the core issue in the
backend, but until that fix is added, the changes in this commit
should prevent the volume-detail commands from resulting in an
error if any invalid transactionStatus objects are returned.