Add fix for unsetting of values in edit SG rules
The SL API for editing security group rules allows you to unset string
values by passing in an empty string (''). The edit_securitygroup_rule()
function skipped over any values that were falsy (because if they
weren't set in the arguments, we don't want to change them).
Unfortunately, this caused values that were an empty string to also get
skipped over and not get unset in the API. So edit needs to check
explicitly for None (don't change) in order to get skipped over in
passing to the API.