Snap for SLCLI
I've created a "snap" package for `slcli`; testing and is working quite well. Snaps are universal packages for Linux distros and libs and software within a snap will not affect those on your system. 

This snap is already up and running and can be installed by running this command from your terminal:

`sudo snap install slcli`